Defer loading of journal items actions menu for improved performance
