+++ app/components/work_packages/activities_tab/journals/item_component.rb
@@ -51,6 +51,10 @@ module WorkPackages
 
         attr_reader :journal, :state, :filter, :grouped_emoji_reactions
 
+        def menu_id
+          ItemComponent::Actions.menu_id(journal)
+        end
+
         def wrapper_uniq_by
           journal.id
         end
@@ -106,69 +110,6 @@ module WorkPackages
 
         def notification_on_details?
           has_unread_notifications? && journal.notes.blank?
-        end
-
-        def allowed_to_edit?
-          journal.editable_by?(User.current)
-        end
-
-        def allowed_to_quote?
-          User.current.allowed_in_project?(:add_work_package_comments, journal.journable.project)
-        end
-
-        def copy_url_action_item(menu)
-          menu.with_item(label: t("button_copy_link_to_clipboard"),
-                         tag: :button,
-                         content_arguments: {
-                           data: {
-                             action: "click->work-packages--activities-tab--item#copyActivityUrlToClipboard"
-                           }
-                         }) do |item|
-            item.with_leading_visual_icon(icon: :copy)
-          end
-        end
-
-        def edit_action_item(menu)
-          menu.with_item(label: edit_action_label,
-                         href: edit_work_package_activity_path(journal.journable, journal, filter:),
-                         content_arguments: {
-                           data: { turbo_stream: true, test_selector: "op-wp-journal-#{journal.id}-edit" }
-                         }) do |item|
-            item.with_leading_visual_icon(icon: :pencil)
-          end
-        end
-
-        def edit_action_label
-          if journal.user == User.current
-            t("js.label_edit_comment")
-          else
-            t("js.label_moderate_comment")
-          end
-        end
-
-        def quote_action_item(menu)
-          menu.with_item(label: t("js.label_quote_comment"),
-                         tag: :button,
-                         content_arguments: {
-                           data: quote_action_data_attributes
-                         }) do |item|
-            item.with_leading_visual_icon(icon: :quote)
-          end
-        end
-
-        def quote_action_data_attributes # rubocop:disable Metrics/AbcSize
-          {
-            test_selector: "op-wp-journal-#{journal.id}-quote",
-            controller: quote_comments_stimulus_controller,
-            action: "click->#{quote_comments_stimulus_controller}#quote:prevent",
-            quote_comments_stimulus_controller("-content-param") => journal.notes,
-            quote_comments_stimulus_controller("-user-id-param") => journal.user_id,
-            quote_comments_stimulus_controller("-user-name-param") => journal.user.name,
-            quote_comments_stimulus_controller("-is-internal-param") => journal.internal?,
-            quote_comments_stimulus_controller("-text-wrote-param") => I18n.t(:text_wrote),
-            quote_comments_stimulus_controller("-#{internal_comment_stimulus_controller}-outlet") => add_comment_component_dom_selector, # rubocop:disable Layout/LineLength
-            quote_comments_stimulus_controller("-#{editor_stimulus_controller}-outlet") => index_component_dom_selector
-          }
         end
       end
     end
