+++ app/components/work_packages/activities_tab/journals/item_component/actions.html.erb
@@ -1,0 +1,41 @@
+<%=
+  render(
+    Primer::Alpha::ActionMenu::List.new(menu_id:)
+  ) do |menu|
+    menu.with_item(
+      label: t("button_copy_link_to_clipboard"),
+      tag: :button,
+      content_arguments: {
+        data: {
+          action: "click->work-packages--activities-tab--item#copyActivityUrlToClipboard"
+        }
+      }
+    ) do |item|
+      item.with_leading_visual_icon(icon: :copy)
+    end
+
+    if editable?
+      menu.with_item(
+        label: edit_action_label,
+        href: edit_work_package_activity_path(journal.journable, journal),
+        content_arguments: {
+          data: { turbo_stream: true, test_selector: "op-wp-journal-#{journal.id}-edit" }
+        }
+      ) do |item|
+        item.with_leading_visual_icon(icon: :pencil)
+      end
+    end
+
+    if quotable?
+      menu.with_item(
+        label: I18n.t("js.label_quote_comment"),
+        tag: :button,
+        content_arguments: {
+          data: quote_action_data_attributes
+        }
+      ) do |item|
+        item.with_leading_visual_icon(icon: :quote)
+      end
+    end
+  end
+%>
